cid-3525: Change the full scan to async method
